requirements.txt
fastapi
uvicorn
numpy
Pillow
face_recognition
python-multipart
